Init metrics.
